+++ src/components/MyMedicalReport/BottomButtons.tsx
@@ -1,4 +1,3 @@
-import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 const BottomButtons = () => {
