+++ src/constants/stageStyle.ts
@@ -1,0 +1,43 @@
+const stageStyleMap = {
+  정상: {
+    border: '1px solid #64DF48',
+    background:
+      'radial-gradient(79.04% 79.48% at 50% 50%, rgba(100, 223, 72, 0.20) 0%, rgba(100, 223, 72, 0.00) 100%)',
+    textColor: '#76E15D',
+    circleFill: 'rgba(118, 225, 93, 0.50)',
+    dashedStroke: '#76E15D',
+  },
+  주의: {
+    border: '1px solid #FF732D',
+    background:
+      'radial-gradient(79.04% 79.48% at 50% 50%, rgba(255, 115, 45, 0.20) 0%, rgba(255, 115, 45, 0.00) 100%)',
+    textColor: '#FF732D',
+    circleFill: 'rgba(255, 115, 45, 0.50)',
+    dashedStroke: '#FF732D',
+  },
+  위험: {
+    border: '1px solid #ED5151',
+    background:
+      'radial-gradient(79.04% 79.48% at 50% 50%, rgba(237, 81, 81, 0.20) 0%, rgba(237, 81, 81, 0.00) 100%)',
+    textColor: '#ED5151',
+    circleFill: 'rgba(237, 81, 81, 0.50)',
+    dashedStroke: '#ED5151',
+  },
+  안심: {
+    border: '1px solid #1D68FF',
+    background:
+      'radial-gradient(79.04% 79.48% at 50% 50%, rgba(29, 104, 255, 0.20) 0%, rgba(29, 104, 255, 0.00) 100%)',
+    textColor: '#1D68FF',
+    circleFill: 'rgba(29, 104, 255, 0.50)',
+    dashedStroke: '#1D68FF',
+  },
+  관심: {
+    border: '1px solid #FFCC00',
+    background:
+      'radial-gradient(79.04% 79.48% at 50% 50%, rgba(255, 204, 0, 0.20) 0%, rgba(255, 204, 0, 0.00) 100%)',
+    textColor: '#FFCC00',
+    circleFill: 'rgba(255, 204, 0, 0.50)',
+    dashedStroke: '#FFCC00',
+  },
+};
+export default stageStyleMap;
